Change the input parameter type: uint32 -> uint256
There is no need to replace uint256 into uint32 because both use 32bytes. Storage variables are packed, but this is not true in memory and calldata.

+++ contracts/StakingPoolV2.sol
@@ -185,7 +185,7 @@ contract StakingPoolV2 is IStakingPoolV2, StakedElyfiToken {
   function initNewPool(
     uint256 rewardPerSecond,
     uint256 startTimestamp,
-    uint32 duration
+    uint256 duration
   ) external override onlyAdmin {
     if (_poolData.isFinished == true) revert Finished();
     (uint256 newRoundStartTimestamp, uint256 newRoundEndTimestamp) = _poolData.initRound(
